+++ sysproc.c
@@ -159,7 +159,7 @@ int sys_getinfo(void) {
 int sys_increment(void) {
     struct sum *suminfo;
     struct info *pinfo;
-    long long *tat;
+    int *tat;
 
     if (argptr(0, (void *) &pinfo, sizeof(*pinfo)) || argptr(1, (void *) &suminfo, sizeof(*suminfo)) ||
         argptr(2, (void *) &tat, sizeof(*tat)) < 0)
